Set left option key to 'Esc+' and right option key to 'Normal'
Esc+ means that the key acts as a Meta key. With the current
configuration, the left option key can be used as a Meta key, for
example, for GNU Readline key bindings that use the Meta key, such as
M-f, M-b, M-Backspace for navigating and editing the command line.

On the other hand, the right option key can be used as a normal option
key (Alt), for example, for typing characters such as @ (Alt-2), #
(Alt-3), { (Alt-ä), or [ (Alt-ü).

Note that the left option key can only be used as a Meta key and NOT as
an option key. Similarly, the right option key can only be used as an
option (Alt) key and NOT as a Meta key.

That means, all Meta operation must be done with the left option key and
all Alt opertions must be done with the right option key. However, this
seems to be the way that I have done it before anyway.